Release v0.16.0

+++ src/aind_data_asset_indexer/codeocean_bucket_indexer.py
@@ -7,25 +7,24 @@ import logging
 import os
 import sys
 import warnings
-from datetime import datetime
-from typing import List, Optional, Union
+from typing import List, Optional
 
 import boto3
 import dask.bag as dask_bag
+import requests
+from aind_data_access_api.document_db import MetadataDbClient
+from aind_data_access_api.utils import get_s3_bucket_and_prefix, paginate_docdb
 from aind_data_schema.core.metadata import ExternalPlatforms
 from codeocean import CodeOcean
 from codeocean.data_asset import DataAssetSearchOrigin, DataAssetSearchParams
 from mypy_boto3_s3 import S3Client
-from pymongo import MongoClient
-from pymongo.operations import UpdateOne
+from requests.adapters import HTTPAdapter
 from urllib3.util import Retry
 
 from aind_data_asset_indexer.models import CodeOceanIndexBucketJobSettings
 from aind_data_asset_indexer.utils import (
     build_metadata_record_from_prefix,
     get_all_processed_codeocean_asset_records,
-    get_s3_bucket_and_prefix,
-    paginate_docdb,
 )
 
 logging.basicConfig(level=os.getenv("LOG_LEVEL", "INFO"))
@@ -51,6 +50,24 @@ class CodeOceanIndexBucketJob:
     def __init__(self, job_settings: CodeOceanIndexBucketJobSettings):
         """Class constructor."""
         self.job_settings = job_settings
+
+    def _create_docdb_client(self) -> MetadataDbClient:
+        """Create a MetadataDbClient with custom retries."""
+        retry = Retry(
+            total=3,
+            backoff_factor=1,
+            status_forcelist=[429, 500, 502, 503, 504],
+            allowed_methods=["GET", "POST", "DELETE"],
+        )
+        adapter = HTTPAdapter(max_retries=retry)
+        session = requests.Session()
+        session.mount("https://", adapter)
+        return MetadataDbClient(
+            host=self.job_settings.doc_db_host,
+            database=self.job_settings.doc_db_db_name,
+            collection=self.job_settings.doc_db_collection_name,
+            session=session,
+        )
 
     @staticmethod
     def _get_external_data_asset_records(
@@ -129,24 +146,23 @@ class CodeOceanIndexBucketJob:
         return new_records
 
     @staticmethod
-    def _get_co_links_from_record(
-        docdb_record: Union[dict, list]
-    ) -> List[str]:
+    def _get_co_links_from_record(docdb_record: dict) -> List[str]:
         """
         Small utility to parse the external_links field of the docdb record.
-        Supports the legacy type.
+        Supports the legacy type for the external_links field.
 
         Parameters
         ----------
-        docdb_record : dict | list
-          The legacy type was a list, while the current version is a dict.
+        docdb_record : dict
+          The legacy external_links type was a list, while the current
+          version is a dict.
 
         Returns
         -------
         List[str]
 
         """
-        external_links = docdb_record.get("external_links", [])
+        external_links = docdb_record.get("external_links", dict())
 
         # Hopefully, ExternalPlatforms.CODEOCEAN doesn't change
         if isinstance(external_links, dict):
@@ -161,7 +177,7 @@ class CodeOceanIndexBucketJob:
         return external_links
 
     def _update_external_links_in_docdb(
-        self, docdb_client: MongoClient, co_client: CodeOcean
+        self, docdb_client: MetadataDbClient, co_client: CodeOcean
     ) -> None:
         """
         This method will:
@@ -172,7 +188,7 @@ class CodeOceanIndexBucketJob:
 
         Parameters
         ----------
-        docdb_client : MongoClient
+        docdb_client : MetadataDbClient
 
         Returns
         -------
@@ -183,16 +199,12 @@ class CodeOceanIndexBucketJob:
         list_of_co_ids_and_locations = self._get_external_data_asset_records(
             co_client=co_client
         )
-        db = docdb_client[self.job_settings.doc_db_db_name]
-        collection = db[self.job_settings.doc_db_collection_name]
         if list_of_co_ids_and_locations is not None:
             co_loc_to_id_map = self._map_external_list_to_dict(
                 list_of_co_ids_and_locations
             )
             pages = paginate_docdb(
-                docdb_client=docdb_client,
-                db_name=self.job_settings.doc_db_db_name,
-                collection_name=self.job_settings.doc_db_collection_name,
+                docdb_api_client=docdb_client,
                 filter_query={
                     "location": {
                         "$not": {
@@ -214,18 +226,16 @@ class CodeOceanIndexBucketJob:
                         else co_loc_to_id_map.get(location)
                     )
                     docdb_rec_id = record["_id"]
-                    if (
-                        external_links is not None
-                        and code_ocean_ids is not None
-                        and code_ocean_ids != set(external_links)
+                    if code_ocean_ids is not None and code_ocean_ids != set(
+                        external_links
                     ):
                         new_external_links = code_ocean_ids
-                    elif external_links is not None and not code_ocean_ids:
+                    elif external_links and not code_ocean_ids:
                         logging.info(
                             f"No code ocean data asset ids found for "
                             f"{location}. Removing external links from record."
                         )
-                        new_external_links = dict()
+                        new_external_links = set()
                     else:
                         new_external_links = None
                     if new_external_links is not None:
@@ -234,32 +244,27 @@ class CodeOceanIndexBucketJob:
                                 list(new_external_links)
                             )
                         }
-                        last_modified = datetime.utcnow().isoformat()
                         records_to_update.append(
-                            UpdateOne(
-                                filter={"_id": docdb_rec_id},
-                                update={
-                                    "$set": {
-                                        "external_links": record_links,
-                                        "last_modified": last_modified,
-                                    }
-                                },
-                                upsert=False,
-                            )
+                            {
+                                "_id": docdb_rec_id,
+                                "external_links": record_links,
+                            }
                         )
                 if len(records_to_update) > 0:
                     logging.info(f"Updating {len(records_to_update)} records")
-                    write_response = collection.bulk_write(
-                        requests=records_to_update
+                    write_responses = (
+                        docdb_client.upsert_list_of_docdb_records(
+                            records=records_to_update
+                        )
                     )
-                    logging.debug(write_response)
+                    logging.debug([r.json() for r in write_responses])
         else:
             logging.error("There was an error retrieving external links!")
 
     def _process_codeocean_record(
         self,
         codeocean_record: dict,
-        docdb_client: MongoClient,
+        docdb_client: MetadataDbClient,
         s3_client: S3Client,
     ):
         """
@@ -271,7 +276,7 @@ class CodeOceanIndexBucketJob:
         Parameters
         ----------
         codeocean_record : dict
-        docdb_client : MongoClient
+        docdb_client : MetadataDbClient
         s3_client : S3Client
 
         """
@@ -292,12 +297,12 @@ class CodeOceanIndexBucketJob:
         )
         if new_metadata_contents is not None:
             logging.info(f"Uploading metadata record for: {location}")
-            db = docdb_client[self.job_settings.doc_db_db_name]
-            collection = db[self.job_settings.doc_db_collection_name]
             # noinspection PyTypeChecker
             json_contents = json.loads(new_metadata_contents)
-            x = collection.insert_one(json_contents)
-            logging.debug(x.inserted_id)
+            response = docdb_client.insert_one_docdb_record(
+                record=json_contents
+            )
+            logging.debug(response.json())
         else:
             logging.warning(
                 f"Unable to build metadata record for: {location}!"
@@ -306,7 +311,7 @@ class CodeOceanIndexBucketJob:
     def _dask_task_to_process_record_list(self, record_list: List[dict]):
         """
         The task to perform within a partition. If n_partitions is set to 20
-        and the outer prefix list had length 1000, then this should process
+        and the outer record list had length 1000, then this should process
         50 code ocean records.
 
         Parameters
@@ -314,31 +319,26 @@ class CodeOceanIndexBucketJob:
         record_list : List[dict]
 
         """
-        # create a s3_client here since dask doesn't serialize it
+        # create clients here since dask doesn't serialize them
         s3_client = boto3.client("s3")
-        doc_db_client = MongoClient(
-            host=self.job_settings.doc_db_host,
-            port=self.job_settings.doc_db_port,
-            retryWrites=False,
-            directConnection=True,
-            username=self.job_settings.doc_db_user_name,
-            password=self.job_settings.doc_db_password.get_secret_value(),
-            authSource="admin",
-        )
-
-        for record in record_list:
-            try:
-                self._process_codeocean_record(
-                    codeocean_record=record,
-                    docdb_client=doc_db_client,
-                    s3_client=s3_client,
-                )
-            except Exception as e:
-                logging.error(
-                    f'Error processing {record.get("location")}: {repr(e)}'
-                )
+        with self._create_docdb_client() as doc_db_client:
+            for record in record_list:
+                try:
+                    self._process_codeocean_record(
+                        codeocean_record=record,
+                        docdb_client=doc_db_client,
+                        s3_client=s3_client,
+                    )
+                except requests.HTTPError as e:
+                    logging.error(
+                        f'Error processing {record.get("location")}: {repr(e)}'
+                        f". Response Body: {e.response.text}"
+                    )
+                except Exception as e:
+                    logging.error(
+                        f'Error processing {record.get("location")}: {repr(e)}'
+                    )
         s3_client.close()
-        doc_db_client.close()
 
     def _process_codeocean_records(self, records: List[dict]):
         """
@@ -361,7 +361,7 @@ class CodeOceanIndexBucketJob:
     def _dask_task_to_delete_record_list(self, record_list: List[str]):
         """
         The task to perform within a partition. If n_partitions is set to 20
-        and the outer prefix list had length 1000, then this should process
+        and the outer record list had length 1000, then this should process
         50 ids.
 
         Parameters
@@ -369,27 +369,16 @@ class CodeOceanIndexBucketJob:
         record_list : List[str]
 
         """
-        # create a s3_client here since dask doesn't serialize it
-        docdb_client = MongoClient(
-            host=self.job_settings.doc_db_host,
-            port=self.job_settings.doc_db_port,
-            retryWrites=False,
-            directConnection=True,
-            username=self.job_settings.doc_db_user_name,
-            password=self.job_settings.doc_db_password.get_secret_value(),
-            authSource="admin",
-        )
-        db = docdb_client[self.job_settings.doc_db_db_name]
-        collection = db[self.job_settings.doc_db_collection_name]
-        try:
-            logging.info(f"Removing {len(record_list)} records")
-            response = collection.delete_many(
-                filter={"_id": {"$in": record_list}}
-            )
-            logging.debug(response.raw_result)
-        except Exception as e:
-            logging.error(f"Error deleting records: {repr(e)}")
-        docdb_client.close()
+        # create a docdb_client here since dask doesn't serialize it
+        with self._create_docdb_client() as docdb_client:
+            try:
+                logging.info(f"Removing {len(record_list)} records")
+                response = docdb_client.delete_many_records(
+                    data_asset_record_ids=record_list
+                )
+                logging.debug(response.json())
+            except Exception as e:
+                logging.error(f"Error deleting records: {repr(e)}")
 
     def _delete_records_from_docdb(self, record_list: List[str]):
         """
@@ -430,38 +419,27 @@ class CodeOceanIndexBucketJob:
         )
         logging.info("Finished scanning through CodeOcean.")
         logging.info("Starting to scan through DocDb.")
-        iterator_docdb_client = MongoClient(
-            host=self.job_settings.doc_db_host,
-            port=self.job_settings.doc_db_port,
-            retryWrites=False,
-            directConnection=True,
-            username=self.job_settings.doc_db_user_name,
-            password=self.job_settings.doc_db_password.get_secret_value(),
-            authSource="admin",
-        )
-        # Use existing client to add external links to fields
-        logging.info("Adding links to records.")
-        self._update_external_links_in_docdb(
-            docdb_client=iterator_docdb_client, co_client=co_client
-        )
-        logging.info("Finished adding links to records")
-        all_docdb_records = dict()
-        docdb_pages = paginate_docdb(
-            db_name=self.job_settings.doc_db_db_name,
-            docdb_client=iterator_docdb_client,
-            collection_name=self.job_settings.doc_db_collection_name,
-            page_size=500,
-            filter_query={
-                "location": {
-                    "$regex": f"^s3://{self.job_settings.s3_bucket}.*"
-                }
-            },
-            projection={"location": 1, "_id": 1},
-        )
-        for page in docdb_pages:
-            for record in page:
-                all_docdb_records[record["location"]] = record["_id"]
-        iterator_docdb_client.close()
+        with self._create_docdb_client() as iterator_docdb_client:
+            # Use existing client to add external links to fields
+            logging.info("Adding links to records.")
+            self._update_external_links_in_docdb(
+                docdb_client=iterator_docdb_client, co_client=co_client
+            )
+            logging.info("Finished adding links to records")
+            all_docdb_records = dict()
+            docdb_pages = paginate_docdb(
+                docdb_api_client=iterator_docdb_client,
+                page_size=500,
+                filter_query={
+                    "location": {
+                        "$regex": f"^s3://{self.job_settings.s3_bucket}.*"
+                    }
+                },
+                projection={"location": 1, "_id": 1},
+            )
+            for page in docdb_pages:
+                for record in page:
+                    all_docdb_records[record["location"]] = record["_id"]
         logging.info("Finished scanning through DocDB.")
         codeocean_locations = set(code_ocean_records.keys())
         docdb_locations = set(all_docdb_records.keys())
@@ -471,13 +449,17 @@ class CodeOceanIndexBucketJob:
             records_to_add.append(code_ocean_records[location])
         for location in docdb_locations - codeocean_locations:
             records_to_delete.append(all_docdb_records[location])
+        logging.info(f"{len(records_to_add)} records to add to DocDB.")
+        logging.info(f"{len(records_to_delete)} records to delete from DocDB.")
 
-        logging.info("Starting to add records to DocDB.")
-        self._process_codeocean_records(records=records_to_add)
-        logging.info("Finished adding records to DocDB.")
-        logging.info("Starting to delete records from DocDB.")
-        self._delete_records_from_docdb(record_list=records_to_delete)
-        logging.info("Finished deleting records from DocDB.")
+        if len(records_to_add) > 0:
+            logging.info("Starting to add records to DocDB.")
+            self._process_codeocean_records(records=records_to_add)
+            logging.info("Finished adding records to DocDB.")
+        if len(records_to_delete) > 0:
+            logging.info("Starting to delete records from DocDB.")
+            self._delete_records_from_docdb(record_list=records_to_delete)
+            logging.info("Finished deleting records from DocDB.")
 
 
 if __name__ == "__main__":
